requirements.txt
django>=5.2
python-dotenv>=1.0
langgraph>=0.2.30
memory-profiler>=0.61
matplotlib>=3.9
openai>=1.40

# Optional (kept for future phases)
gitpython>=3.1
psutil>=5.9
